[SELC-4603] Resolve conflit after merge main
